Reduced trait requirements

+++ src/exact_hyper_spheres_sketch.rs
@@ -6,7 +6,7 @@ use std::collections::HashSet;
 
 impl<I, C> SetLike<C> for HashSet<I>
 where
-    C: Primitive<usize>,
+    C: Primitive<f32>,
     I: Eq + Hash,
 {
     fn get_estimated_union_cardinality(
@@ -15,14 +15,14 @@ where
         other: &Self,
         other_cardinality: C,
     ) -> EstimatedUnionCardinalities<C> {
-        let union_cardinality = C::reverse(self.union(other).count());
+        let union_cardinality = C::reverse(self.union(other).count() as f32);
 
         EstimatedUnionCardinalities::from((self_cardinality, other_cardinality, union_cardinality))
     }
 
     fn get_cardinality(&self) -> C {
-        C::reverse(self.len())
+        C::reverse(self.len() as f32)
     }
 }
 
-impl<C: Eq + Hash, I: Primitive<usize> + One> HyperSpheresSketch<I> for HashSet<C> {}
+impl<C: Eq + Hash, I: Primitive<f32> + One> HyperSpheresSketch<I> for HashSet<C> {}
